requirements.txt
absl-py==2.1.0
accelerate==0.33.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.4.0
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.13.1
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
six==1.16.0
sympy==1.13.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.5.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0
